week10/generate_property_geojson/requirements.txt
python-dotenv
google-cloud-bigquery
google-cloud-storage